Add test case for not overwriting locks acquired by another user
